% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HS102_votes}
\alias{HS102_votes}
\title{House votes data for the 102nd Congress}
\format{
A data frame with the same structure as HS91_votes.
}
\description{
Dataset containing voting information for the 102nd Congress (1991-1993).
}
\seealso{
\code{\link{HS91_votes}} for a detailed description of the variables.
}
