% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HS107_votes}
\alias{HS107_votes}
\title{House roll calls for the 107th Congress}
\format{
A data frame with the same structure as HS91_votes.
}
\description{
Data set containing voting information for the 107th Congress (2001-2003).
}
\seealso{
\code{\link{HS91_votes}} for a detailed description of the variables.
}
