% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HS117_votes}
\alias{HS117_votes}
\title{House roll calls for the 117th Congress}
\format{
A data frame with the same structure as HS91_votes.
}
\description{
Data set containing voting information for the 117th Congress (2021-2023).
}
\seealso{
\code{\link{HS91_votes}} for a detailed description of the variables.
}
