% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HS91_votes}
\alias{HS91_votes}
\title{House votes data for the 91st Congress}
\format{
A data frame with the following variables:
\describe{
  \item{congress}{Integer 1+. The number of the congress for this vote.}
  \item{chamber}{Character. "House", "Senate", or "President". The chamber in which the vote occurred.}
  \item{rollnumber}{Integer 1+. Starts from 1 in the first rollcall of each congress.}
  \item{icpsr}{Integer 1-99999. ID code which identifies the member.}
  \item{cast_code}{Integer 0-9. Indicator of how the member voted.
    \itemize{
      \item 0: Not a member of the chamber when this vote was taken
      \item 1: Yea
      \item 2: Paired Yea
      \item 3: Announced Yea
      \item 4: Announced Nay
      \item 5: Paired Nay
      \item 6: Nay
      \item 7: Present (some Congresses)
      \item 8: Present (some Congresses)
      \item 9: Not Voting (Abstention)
    }
  }
  \item{prob}{Numeric. Estimated probability, based on NOMINATE, of the member making the vote as recorded.}
}
}
\description{
Dataset containing voting information for the 91st Congress (1969-1971).
}
