% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{loading}
\alias{loading}
\title{Loading of U.S. House data sets.}
\usage{
loading(user_input)
}
\arguments{
\item{user_input}{Number equalling either 91, 97, 102, 107 or 117.}
}
\value{
A list object 'house' containing members and their votes of a chosen Congress. This can be done repetitively, for multiple congresses.
}
\description{
Loads United States House Representatives and their votes into a house list, based on the user's chosen Congress.
}
\examples{
loading("91")
loading("97")
}
