% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phillomenal-package.R
\docType{package}
\name{phillomenal}
\alias{phillomenal-package}
\alias{phillomenal}
\title{phillomenal: A package for analyzing U.S. House voting data}
\description{
This package includes data sets of all United States Congress members ("HSall_members") and their votes of selected Congresses ("HS_votes"),
namely 91, 97, 102, 107 and 117.
}
