% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{wrangling}
\alias{wrangling}
\title{Wrangling of U.S. House data sets}
\usage{
wrangling(user_input)
}
\arguments{
\item{user_input}{Number equalling either 91, 97, 102, 107 or 117.}
}
\value{
An updated votes_id, pairs_id which displays all possible pairs of representatives in a given Congress.
}
\description{
Wrangles the objects created in loading(), namely votes_id and members_id, by merging select columns of both objects. Specifically,
votes_id is being enhanced by information about each representative in order to allow the creation of pairs_id for the chosen Congress.
}
\examples{
wrangling("91")
wrangling("97")
}
